Add Constants Class

+++ ITSEP4A20App/app/src/main/java/com/example/it_sep4_a20_app/ui/preferences/temperaturepreference/TemperaturePreferenceFragment.java
@@ -45,7 +45,7 @@ public class TemperaturePreferenceFragment extends PreferenceFragmentCompat
             boolean valid = numberCheck(newValue);
             if(valid)
             {
-                mViewModel.setTemperatureSetPoint((Integer.parseInt(newValue.toString())));
+                mViewModel.setTemperatureSetPoint((Double.parseDouble(newValue.toString())));
             }
             return valid;
         });
